requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
psycopg2

# ----------------------
# Production Server
# ----------------------
gunicorn